Moved memberlist and raft packages to internal folder
