+++ pkg/echovault/echovault.go
@@ -21,10 +21,10 @@ import (
 	"errors"
 	"fmt"
 	"github.com/echovault/echovault/internal"
+	"github.com/echovault/echovault/internal/memberlist"
+	"github.com/echovault/echovault/internal/raft"
 	"github.com/echovault/echovault/pkg/aof"
 	"github.com/echovault/echovault/pkg/eviction"
-	"github.com/echovault/echovault/pkg/memberlist"
-	"github.com/echovault/echovault/pkg/raft"
 	"github.com/echovault/echovault/pkg/snapshot"
 	"github.com/echovault/echovault/pkg/utils"
 	"io"
